Fixed "includes" URL bug and adds "url" query param.
The bug was that if an innocent website had a blocked URL's path anywhere in it (e.g. "example.com/news/twitter.com-is-a-site") then it would get blocked. We now just check the host portion to see if it matches the block list.

The URL query parameter allows us to remember the original site on "blocked.html" so that we can redirect back to there when focus mode is disabled.